test: 인수테스트 작성-사용자의 예약을 관리자가 체크인 상태로 변경한다.
