+++ src/test/java/com/camping/CucumberTest.java
@@ -1,4 +1,4 @@
-package features;
+package com.camping;
 
 import org.junit.platform.suite.api.SelectClasspathResource;
 import org.junit.platform.suite.api.Suite;
