+++ src/test/resources/features/reservation.feature
@@ -4,3 +4,8 @@ Feature: 관리자 예약 관리 기능
     Given 사용자가 예약을 했다
     When 관리자가 예약을 취소했다
     Then 예약은 취소 상태다
+
+  Scenario: 사용자의 예약을 관리자가 체크인 상태로 변경한다.
+    Given 사용자가 예약을 했다
+    When 관리자가 예약을 체크인 상태로 변경을 요청한다
+    Then 예약은 체크인 상태다
